+++ great_expectations/datasource/types/__init__.py
@@ -1,0 +1,23 @@
+import enum
+
+from .batch_kwargs import *
+from .batch_spec import *
+
+
+# noinspection SpellCheckingInspection
+class DatasourceTypes(enum.Enum):
+    PANDAS = "pandas"
+    SPARK = "spark"
+    SQL = "sql"
+    # TODO DBT = "dbt"
+
+
+# noinspection SpellCheckingInspection
+class SupportedDatabases(enum.Enum):
+    MYSQL = "MySQL"
+    POSTGRES = "Postgres"
+    REDSHIFT = "Redshift"
+    SNOWFLAKE = "Snowflake"
+    BIGQUERY = "BigQuery"
+    OTHER = "other - Do you have a working SQLAlchemy connection string?"
+    # TODO MSSQL