+++ great_expectations/execution_engine/sparkdf_execution_engine.py
@@ -7,12 +7,12 @@ from typing import Any, Callable, Dict, Iterable, Optional, Tuple, Union
 
 from great_expectations.core.batch import BatchMarkers, BatchSpec
 from great_expectations.core.id_dict import IDDict
-from great_expectations.exceptions import exceptions as ge_exceptions
-from great_expectations.execution_environment.types.batch_spec import (
+from great_expectations.datasource.types.batch_spec import (
     PathBatchSpec,
     RuntimeDataBatchSpec,
     S3BatchSpec,
 )
+from great_expectations.exceptions import exceptions as ge_exceptions
 
 from ..exceptions import (
     BatchKwargsError,
@@ -40,6 +40,15 @@ try:
         StructField,
         StructType,
     )
+
+    class SparkDFBatchData(DataFrame):
+        def __init__(self, df):
+            super(self.__class__, self).__init__(df._jdf, df.sql_ctx)
+
+        def row_count(self):
+            return self.count()
+
+
 except ImportError:
     pyspark = None
     SparkSession = None
@@ -52,43 +61,12 @@ except ImportError:
     StringType = (None,)
     DateType = (None,)
     BooleanType = (None,)
+
+    SparkDFBatchData = None
+
     logger.debug(
         "Unable to load pyspark; install optional spark dependency for support."
     )
-
-
-class SparkDFBatchData:
-    def __init__(
-        self,
-        dataframe: DataFrame = None,
-        dataframe_dict: Dict[str, DataFrame] = None,
-        default_table_name=None,
-    ):
-        assert (
-            dataframe is not None or dataframe_dict is not None
-        ), "dataframe or dataframe_dict is required"
-        assert (
-            not dataframe and dataframe_dict
-        ), "dataframe and dataframe_dict may not both be specified"
-
-        if dataframe is not None:
-            dataframe_dict = {"": dataframe}
-            default_table_name = ""
-
-        self._dataframe_dict = dataframe_dict
-        self._default_table_name = default_table_name
-
-    @property
-    def default(self):
-        if self._default_table_name in self._dataframe_dict:
-            return self._dataframe_dict[self._default_table_name]
-        return None
-
-    def __getattr__(self, item):
-        return self._dataframe_dict.get(item)
-
-    def __getitem__(self, item):
-        return self._dataframe_dict.get(item)
 
 
 class SparkDFExecutionEngine(ExecutionEngine):
@@ -241,8 +219,9 @@ This class holds an attribute `spark_df` which is a spark.sql.DataFrame.
             )
 
         batch_data = self._apply_splitting_and_sampling_methods(batch_spec, batch_data)
+        typed_batch_data = SparkDFBatchData(batch_data)
 
-        return batch_data, batch_markers
+        return typed_batch_data, batch_markers
 
     def _apply_splitting_and_sampling_methods(self, batch_spec, batch_data):
         if batch_spec.get("splitter_method"):