+++ tests/datasource/test_sparkdf_datasource.py
@@ -9,8 +9,8 @@ from great_expectations.core.batch import Batch
 from great_expectations.core.expectation_suite import ExpectationSuite
 from great_expectations.dataset import SparkDFDataset
 from great_expectations.datasource import SparkDFDatasource
+from great_expectations.datasource.types import InMemoryBatchKwargs
 from great_expectations.exceptions import BatchKwargsError
-from great_expectations.execution_environment.types import InMemoryBatchKwargs
 from great_expectations.util import is_library_loadable
 from great_expectations.validator.validator import BridgeValidator, Validator