+++ tests/execution_engine/test_pandas_execution_engine.py
@@ -10,19 +10,19 @@ from moto import mock_s3
 
 import great_expectations.exceptions.exceptions as ge_exceptions
 from great_expectations.core.batch import Batch
+from great_expectations.datasource.data_connector import (
+    ConfiguredAssetS3DataConnector,
+    InferredAssetS3DataConnector,
+)
+from great_expectations.datasource.types.batch_spec import (
+    PathBatchSpec,
+    RuntimeDataBatchSpec,
+    S3BatchSpec,
+)
 from great_expectations.exceptions.metric_exceptions import MetricProviderError
 from great_expectations.execution_engine.execution_engine import MetricDomainTypes
 from great_expectations.execution_engine.pandas_execution_engine import (
     PandasExecutionEngine,
-)
-from great_expectations.execution_environment.data_connector import (
-    ConfiguredAssetS3DataConnector,
-    InferredAssetS3DataConnector,
-)
-from great_expectations.execution_environment.types.batch_spec import (
-    PathBatchSpec,
-    RuntimeDataBatchSpec,
-    S3BatchSpec,
 )
 from great_expectations.validator.validation_graph import MetricConfiguration
 
@@ -318,7 +318,7 @@ def test_get_batch_with_split_on_whole_table_s3_with_configured_asset_s3_data_co
 
     my_data_connector = ConfiguredAssetS3DataConnector(
         name="my_data_connector",
-        execution_environment_name="FAKE_EXECUTION_ENVIRONMENT_NAME",
+        datasource_name="FAKE_DATASOURCE_NAME",
         default_regex={"pattern": "alpha-(.*)\\.csv", "group_names": ["index"],},
         bucket=bucket,
         prefix="",