+++ tests/validator/test_validator.py
@@ -135,17 +135,17 @@ def test_populate_dependencies_with_incorrect_metric_name():
     assert isinstance(graph, MetricProviderError)
 
 
-def test_graph_validate(basic_execution_environment):
+def test_graph_validate(basic_datasource):
     df = pd.DataFrame({"a": [1, 5, 22, 3, 5, 10], "b": [1, 2, 3, 4, 5, None]})
     expectationConfiguration = ExpectationConfiguration(
         expectation_type="expect_column_value_z_scores_to_be_less_than",
         kwargs={"column": "b", "mostly": 0.9, "threshold": 4, "double_sided": True,},
     )
 
-    batch = basic_execution_environment.get_single_batch_from_batch_request(
+    batch = basic_datasource.get_single_batch_from_batch_request(
         BatchRequest(
             **{
-                "execution_environment_name": "my_execution_environment",
+                "datasource_name": "my_datasource",
                 "data_connector_name": "test_runtime_data_connector",
                 "batch_data": df,
                 "partition_request": PartitionRequest(
@@ -184,7 +184,7 @@ def test_graph_validate(basic_execution_environment):
 
 
 # this might indicate that we need to validate configuration a little more strictly prior to actually validating
-def test_graph_validate_with_bad_config(basic_execution_environment):
+def test_graph_validate_with_bad_config(basic_datasource):
     df = pd.DataFrame({"a": [1, 5, 22, 3, 5, 10], "b": [1, 2, 3, 4, 5, None]})
     expectationConfiguration = ExpectationConfiguration(
         expectation_type="expect_column_max_to_be_between",
@@ -192,10 +192,10 @@ def test_graph_validate_with_bad_config(basic_execution_environment):
     )
     expectation = ExpectColumnMaxToBeBetween(expectationConfiguration)
 
-    batch = basic_execution_environment.get_single_batch_from_batch_request(
+    batch = basic_datasource.get_single_batch_from_batch_request(
         BatchRequest(
             **{
-                "execution_environment_name": "my_execution_environment",
+                "datasource_name": "my_datasource",
                 "data_connector_name": "test_runtime_data_connector",
                 "batch_data": df,
                 "partition_request": PartitionRequest(
@@ -221,7 +221,7 @@ def test_graph_validate_with_bad_config(basic_execution_environment):
 
 
 # Tests that runtime configuration actually works during graph validation
-def test_graph_validate_with_runtime_config(basic_execution_environment):
+def test_graph_validate_with_runtime_config(basic_datasource):
     df = pd.DataFrame(
         {"a": [1, 5, 22, 3, 5, 10, 2, 3], "b": [97, 332, 3, 4, 5, 6, 7, None]}
     )
@@ -231,10 +231,10 @@ def test_graph_validate_with_runtime_config(basic_execution_environment):
     )
     expectation = ExpectColumnValueZScoresToBeLessThan(expectationConfiguration)
 
-    batch = basic_execution_environment.get_single_batch_from_batch_request(
+    batch = basic_datasource.get_single_batch_from_batch_request(
         BatchRequest(
             **{
-                "execution_environment_name": "my_execution_environment",
+                "datasource_name": "my_datasource",
                 "data_connector_name": "test_runtime_data_connector",
                 "batch_data": df,
                 "partition_request": PartitionRequest(